Refactor current modules dirs
