+++ device/vulkan/CMakeLists.txt
@@ -1,0 +1,11 @@
+cmake_minimum_required(VERSION 3.11)
+project(VulkanDevice)
+
+target_sources(VulkanDevice
+    PUBLIC
+        include/VulkanRenderDevice.h
+    PRIVATE
+        sources/VulkanRenderDevice.cpp
+)
+
+add_library(VulkanDevice STATIC)
