+++ device/vulkan/sources/VulkanRenderDevice.cpp
@@ -1,0 +1,5 @@
+//
+// Created by Egor Orachyov on 2019-11-02.
+//
+
+#include "vulkan/include/VulkanRenderDevice.h"
