--- vulkan/CMakeLists.txt
@@ -1,8 +1,0 @@
-cmake_minimum_required(VERSION 3.11)
-project(VulkanDevice)
-
-set(SOURCES
-
-)
-
-# add_library(VulkanDevice STATIC ${SOURCES})
